backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
python-dotenv==1.0.1
openai>=1.76.0
supabase==2.10.0
httpx>=0.27.0,<0.28.0
numpy==2.1.3
scikit-learn==1.5.2
pydantic==2.10.3
pydantic-settings==2.6.1
elevenlabs==1.9.0
google-genai==1.0.0
